refactor: 원시 값을 포장하는 클래스들을 model 패키지로 이동
