+++ src/main/java/lotterymachine/model/WinningLottery.java
@@ -1,8 +1,5 @@
 package lotterymachine.model;
 
-import lotterymachine.dto.Count;
-import lotterymachine.dto.Money;
-
 import java.util.Arrays;
 import java.util.EnumMap;
 import java.util.Map;
